edit category done

+++ resources/views/frontend/mySuggestion.blade.php
@@ -57,7 +57,7 @@
                                                         <li>
                                                             <a href="/ubah-usulan/{{$d->id}}" class="btn"><i class="fa fa-pencil-square-o" aria-hidden="true"></i> Ubah</a></li>
                                                         <li>
-                                                            <a class="btn" href="/hapus-usulan/{{$d->id}}"><i class="fa fa-trash-o" aria-hidden="true"></i> Hapus</a>
+                                                            <a class="btn" href="/hapus-usulan/{{$d->id}}" onclick="return confirm('Apakah Anda yakin ingin menghapus data usulan?')"><i class="fa fa-trash-o" aria-hidden="true"></i> Hapus</a>
                                                         </li>
                                                     @endif
                                                 </ul>
@@ -93,7 +93,7 @@
                                             <p>Tanggal Pengusulan: {{ Carbon\Carbon::parse($d->date)->isoFormat('D MMMM Y') }}</p>
                                             @if($d->status == 'proses review')
                                                 <a href="/ubah-usulan/{{$d->id}}" class="card-link">Ubah</a>
-                                                <a href="/hapus-usulan/{{$d->id}}" class="card-link">Hapus</a>
+                                                <a href="/hapus-usulan/{{$d->id}}" onclick="return confirm('Apakah Anda yakin ingin menghapus data usulan?')" class="card-link">Hapus</a>
                                             @endif
                                             </div>
                                         </div>
